refactor: separate commands and payloads into dedicated modules
- Move command-related code to src/commands/ directory
- Move payload-related code to src/payload/ directory
- Remove command_id dependency from HsesPayload trait
- Add VariableCommandId trait for variable command ID management
- Update client layer to use new trait constraints
- Fix all tests and maintain backward compatibility

This refactoring improves code organization and removes circular dependencies
between payload and command layers.
