+++ moto-hses-client/src/lib.rs
@@ -13,5 +13,5 @@ pub use types::{ClientConfig, ClientError, HsesClient};
 
 // Re-export protocol types that are commonly used
 pub use moto_hses_proto::{
-    Alarm, ControlGroupPositionType, ExecutingJobInfo, Position, Status, TextEncoding, VariableType,
+    Alarm, ControlGroupPositionType, ExecutingJobInfo, HsesPayload, Position, Status, TextEncoding,
 };
