+++ moto-hses-mock/examples/mock_basic_usage.rs
@@ -37,8 +37,8 @@ async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
     info!("  0x87 - Job select");
 
     // Add some test alarms
-    server.add_test_alarm(proto::alarm::test_alarms::servo_error()).await;
-    server.add_test_alarm(proto::alarm::test_alarms::emergency_stop()).await;
+    server.add_test_alarm(proto::payload::alarm::test_alarms::servo_error()).await;
+    server.add_test_alarm(proto::payload::alarm::test_alarms::emergency_stop()).await;
 
     // Set some test variables
     server.set_variable(10, vec![0x42, 0x00, 0x00, 0x00]).await; // D010 = 66
