+++ moto-hses-mock/src/handlers/alarm.rs
@@ -1,9 +1,9 @@
 //! Alarm-related command handlers
 
-use super::CommandHandler;
-use crate::state::MockState;
-use moto_hses_proto as proto;
-use moto_hses_proto::alarm::Alarm;
+use crate::{CommandHandler, state::MockState};
+use moto_hses_proto::{
+    Alarm, HsesRequestMessage, ProtocolError, ReadAlarmData, ReadAlarmHistory, encoding_utils,
+};
 
 /// Common helper function to handle alarm attribute reading based on service type
 fn handle_alarm_service_request(
@@ -11,7 +11,7 @@ fn handle_alarm_service_request(
     service: u8,
     attribute: u8,
     state: &MockState,
-) -> Result<Vec<u8>, proto::ProtocolError> {
+) -> Result<Vec<u8>, ProtocolError> {
     match service {
         0x01 => {
             // Service = 0x01 (Get_Attribute_All) - Return complete alarm data (60 bytes)
@@ -45,8 +45,7 @@ fn get_alarm_attribute_data(alarm: &Alarm, attribute: u8, state: &MockState) -> 
         }
         4 => {
             // Alarm time (16 bytes)
-            let time_bytes =
-                moto_hses_proto::encoding_utils::encode_string(&alarm.time, state.text_encoding);
+            let time_bytes = encoding_utils::encode_string(&alarm.time, state.text_encoding);
             let mut padded_time = vec![0u8; 16];
             padded_time[..time_bytes.len().min(16)]
                 .copy_from_slice(&time_bytes[..time_bytes.len().min(16)]);
@@ -54,8 +53,7 @@ fn get_alarm_attribute_data(alarm: &Alarm, attribute: u8, state: &MockState) -> 
         }
         5 => {
             // Alarm name (32 bytes)
-            let name_bytes =
-                moto_hses_proto::encoding_utils::encode_string(&alarm.name, state.text_encoding);
+            let name_bytes = encoding_utils::encode_string(&alarm.name, state.text_encoding);
             let mut padded_name = vec![0u8; 32];
             padded_name[..name_bytes.len().min(32)]
                 .copy_from_slice(&name_bytes[..name_bytes.len().min(32)]);
@@ -74,24 +72,24 @@ pub struct AlarmDataHandler;
 impl CommandHandler for AlarmDataHandler {
     fn handle(
         &self,
-        message: &proto::HsesRequestMessage,
+        message: &HsesRequestMessage,
         state: &mut MockState,
-    ) -> Result<Vec<u8>, proto::ProtocolError> {
+    ) -> Result<Vec<u8>, ProtocolError> {
         let instance = message.sub_header.instance;
         let attribute = message.sub_header.attribute;
         let service = message.sub_header.service;
 
         // Create ReadAlarmData command to validate instance and attribute
-        let alarm_data_cmd = proto::alarm::ReadAlarmData::new(instance, attribute);
+        let alarm_data_cmd = ReadAlarmData::new(instance, attribute);
 
         // Validate instance range
         if !alarm_data_cmd.is_valid_instance() {
-            return Err(proto::ProtocolError::InvalidCommand);
+            return Err(ProtocolError::InvalidCommand);
         }
 
         // Validate attribute range
         if !alarm_data_cmd.is_valid_attribute() {
-            return Err(proto::ProtocolError::InvalidAttribute);
+            return Err(ProtocolError::InvalidAttribute);
         }
 
         let instance_usize = instance as usize;
@@ -111,19 +109,19 @@ pub struct AlarmInfoHandler;
 impl CommandHandler for AlarmInfoHandler {
     fn handle(
         &self,
-        message: &proto::HsesRequestMessage,
+        message: &HsesRequestMessage,
         state: &mut MockState,
-    ) -> Result<Vec<u8>, proto::ProtocolError> {
+    ) -> Result<Vec<u8>, ProtocolError> {
         let instance = message.sub_header.instance;
         let attribute = message.sub_header.attribute;
         let service = message.sub_header.service;
 
         // Create ReadAlarmHistory command to validate instance
-        let alarm_history_cmd = proto::alarm::ReadAlarmHistory::new(instance, attribute);
+        let alarm_history_cmd = ReadAlarmHistory::new(instance, attribute);
 
         // Validate instance range
         if !alarm_history_cmd.is_valid_instance() {
-            return Err(proto::ProtocolError::InvalidCommand);
+            return Err(ProtocolError::InvalidCommand);
         }
 
         let category = alarm_history_cmd.get_alarm_category();
@@ -146,9 +144,9 @@ pub struct AlarmResetHandler;
 impl CommandHandler for AlarmResetHandler {
     fn handle(
         &self,
-        message: &proto::HsesRequestMessage,
+        message: &HsesRequestMessage,
         state: &mut MockState,
-    ) -> Result<Vec<u8>, proto::ProtocolError> {
+    ) -> Result<Vec<u8>, ProtocolError> {
         let reset_type = message.sub_header.instance;
 
         match reset_type {
