+++ moto-hses-proto/src/lib.rs
@@ -1,38 +1,29 @@
 //! moto-hses-proto - HSES (High Speed Ethernet Server) protocol implementation
 
-pub mod alarm;
 pub mod commands;
 pub mod constants;
-pub mod cycle_mode;
 pub mod encoding;
 pub mod encoding_utils;
 pub mod error;
-pub mod file;
-pub mod job;
 pub mod message;
-pub mod position;
-pub mod status;
-pub mod variables;
+pub mod payload;
 
 // Re-export commonly used items for convenience
-pub use alarm::{Alarm, AlarmAttribute, ReadAlarmData};
 pub use commands::{
-    Command, Division, HoldServoControl, HoldServoType, HoldServoValue, ReadCurrentPosition,
-    ReadIo, ReadRegister, ReadStatus, ReadStatusData1, ReadStatusData2, ReadVar, Service,
-    VariableType, WriteIo, WriteRegister, WriteVar,
+    AlarmAttribute, AlarmReset, Command, CycleMode, CycleModeSwitchingCommand, DeleteFile,
+    Division, HoldServoControl, HoldServoType, HoldServoValue, ReadAlarmData, ReadAlarmHistory,
+    ReadCurrentPosition, ReadExecutingJobInfo, ReadFileList, ReadIo, ReadRegister, ReadStatus,
+    ReadStatusData1, ReadStatusData2, ReadVar, ReceiveFile, SendFile, Service, VariableCommandId,
+    WriteIo, WriteRegister, WriteVar,
 };
 pub use constants::{FILE_CONTROL_PORT, ROBOT_CONTROL_PORT};
-pub use cycle_mode::{CycleMode, CycleModeSwitchingCommand};
 pub use encoding::TextEncoding;
 pub use error::ProtocolError;
-pub use file::response::{parse_file_content, parse_file_list};
-pub use file::{DeleteFile, ReadFileList, ReceiveFile, SendFile};
-pub use job::{ExecutingJobInfo, JobInfoAttribute, ReadExecutingJobInfo, TaskType};
 pub use message::{
     HsesCommonHeader, HsesRequestMessage, HsesRequestSubHeader, HsesResponseMessage,
     HsesResponseSubHeader,
 };
-pub use position::{
-    CartesianPosition, ControlGroupPositionType, CoordinateSystemType, Position, PulsePosition,
+pub use payload::{
+    Alarm, CartesianPosition, ControlGroupPositionType, CoordinateSystemType, ExecutingJobInfo,
+    HsesPayload, Position, PulsePosition, Status, StatusData1, StatusData2,
 };
-pub use status::{Status, StatusData1, StatusData2};
